requirements.txt
streamlit>=1.51.0
google-genai>=1.50.0
Pillow>=12.0.0
